Split TerminalOutputDevice.cs into focused partial files (#9191)
Splits the 801-line TerminalOutputDevice.cs into four focused partial-class files (each under 300 lines):

- TerminalOutputDevice.cs: type declaration, fields, constructor, identity members, Dispose, HandleProcessRoleAsync.
- TerminalOutputDevice.Initialization.cs: InitializeAsync, option-parsing helpers, AnsiOverride enum, GetShortArchitecture.
- TerminalOutputDevice.SessionLifecycle.cs: banner and session start/end display methods plus logging helpers.
- TerminalOutputDevice.DataConsumption.cs: DisplayAsync and ConsumeAsync.

Pure cut-and-paste; no behavioral changes.
